fix: ensure grammars are downloaded during index --reset
When resetting an index, pre-download all supported tree-sitter grammars
to ensure the index is restored to a known good state. This prevents
"Missing grammar" warnings during re-indexing.

Adds ensureGrammarsDownloaded() to setup-helpers.ts as a shared helper.
